+++ examples/02_livejournal.py
@@ -18,7 +18,7 @@ from lxml import html
 from pomp.core.base import BaseCrawler
 from pomp.contrib.pipelines import CsvPipeline
 from pomp.core.base import BasePipeline, BaseDownloaderMiddleware
-from pomp.contrib import UrllibHttpRequest
+from pomp.contrib.urllibtools import UrllibHttpRequest
 from pomp.core.item import Item, Field