+++ tests/test_contrib_urllib.py
@@ -2,7 +2,7 @@ import logging
 from nose.tools import assert_set_equal, assert_equal
 from pomp.core.base import BaseCrawler, BaseDownloaderMiddleware
 from pomp.core.engine import Pomp
-from pomp.contrib import SimpleDownloader, ThreadedDownloader
+from pomp.contrib.urllibtools import UrllibDownloader, ThreadedDownloader
 
 from mockserver import HttpServer, make_sitemap
 from tools import DummyCrawler
@@ -73,7 +73,7 @@ class TestContribUrllib(object):
 
         catch_exception_middleware = CatchException()
         pomp = Pomp(
-            downloader=SimpleDownloader(
+            downloader=UrllibDownloader(
                 middlewares=[catch_exception_middleware]),
             pipelines=[],
         )