+++ CHANGES.rst
@@ -9,6 +9,8 @@ Version 0.1.0
 
 Release date to be decided.
 
+- rename SimpleDownloader to UrllibDownloader
+- urllib code now in contrib/urllibtools.py
 - ENTRY_URL renamed to ENTRY_REQUESTS
 - next_url renamed to next_requests
 - async support